[add] -b -- Force to open a file or folder in the best window
It is default behavior for open file or file from command line.
However, after the -r repair, there was a need for the -b option.
It is useful when a file or folder opens in a new window by default.

The -r option can be used with the -r or -n options:

* `-b -n` -- open file or folder in new window if not find best windows for this.
* `-b -r` -- open file or folder in last active window if not find best windows for this.